package/test_requirements.txt
-r requirements.txt

kedro >=0.18.0
kedro-datasets[pandas.ParquetDataset, pandas.CSVDataset, pandas.ExcelDataset, plotly.JSONDataset]>=2.1.0
kedro-telemetry>=0.1.1  # for testing telemetry integration
bandit~=1.7
behave~=1.2
black~=23.3
boto3~=1.34
flake8~=7.1
isort~=5.11
matplotlib~=3.9
mypy~=1.11
moto~=5.0.9
psutil==5.9.6  # same as Kedro for now
pylint~=3.2
pylint-pydantic>=0.3.0
pytest~=8.3
pytest-asyncio~=0.21
pytest-mock~=3.14
pytest-cov~=5.0
sqlalchemy-stubs~=0.4
strawberry-graphql[cli]>=0.99.0, <1.0
trufflehog~=2.2
httpx~=0.27.0

# mypy
types-aiofiles==0.1.3
types-cachetools==0.1.6
types-click==0.1.14
types-futures==0.1.3
types-Jinja2==2.11.2
types-orjson==0.1.0
types-pkg-resources==0.1.2
types-protobuf==0.1.10
types-PyYAML==0.1.5
types-requests==0.1.8
types-toml==0.1.1
types-ujson==0.1.0

#shareableviz
s3fs>=2021.4
adlfs>=2021.4
gcsfs>=2021.4
